Refactor lnurl to a route
