+++ src/lib/i18n/en.json
@@ -16,7 +16,8 @@
     "settings_currency": "Currency Settings",
     "settings_language": "Language Settings",
     "settings_unit": "Unit Settings",
-    "settings_logs": "Debug Logs"
+    "settings_logs": "Debug Logs",
+    "lnurl": "LNURL"
   },
   "inputs": {
     "destination": {
