+++ src/routes/lnurl/components/pay.svelte
@@ -1,0 +1,2 @@
+<script lang="ts">
+</script>
