+++ src/routes/lnurl/components/withdraw.svelte
@@ -1,0 +1,2 @@
+<script lang="ts">
+</script>
