+++ src/routes/scan/+page.svelte
@@ -14,12 +14,9 @@
   import { convertValue } from '$lib/conversion'
   import lightning from '$lib/lightning'
   import Amount from '$lib/components/Amount.svelte'
-  import LnUrlModal from '$lib/lnurl/modal.svelte'
 
   let requesting = false
   let errorMsg = ''
-
-  let lnurl = ''
 
   const sendPayment$ = new SvelteSubject<
     Pick<Payment, 'bolt11' | 'description' | 'value'> & {
@@ -52,7 +49,7 @@
 
     // check if lnurl
     if (invoice.toLowerCase().startsWith('lnurl')) {
-      lnurl = invoice
+      goto(`/lnurl?lnurl=${invoice}`)
       return
     }
 
@@ -135,11 +132,6 @@
     previousSlide = slide
     slide = i
   }
-
-  function closeLnurlModal() {
-    lnurl = ''
-    goto('/')
-  }
 </script>
 
 <svelte:head>
@@ -188,7 +180,3 @@
 <div class="absolute bottom-4">
   <ErrorMsg bind:message={errorMsg} />
 </div>
-
-{#if lnurl}
-  <LnUrlModal {lnurl} close={closeLnurlModal} />
-{/if}
